+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2021-03-20
+
+### Updated
+
+* Config file name changed from ottosocial.yml|json|toml to config.yml|json|toml
+* Possibility to add config file into the $XDG_CONFIG_HOME directory
+* The file in the configuration path is now called filepath
+
 ## [0.2.0] - 2021-01-09
 
 ### Added